# WCSAxes: pre-compute the coord range to eliminate redundant transformations when drawing ticks/gridlines