@astrofrog 2024-05-01
Nice! How much is the overall performance improvement with this for solar data?