@pllim 2024-05-01
Is this performance covered by https://github.com/astropy/astropy-benchmarks/blob/main/benchmarks/visualization/wcsaxes.py ? If so, can slap a `benchmark` label and let it run.