@astrofrog 2024-05-01
We might as well try that as we should at least see a small improvement even for non solar coordinates!