@pllim 2024-05-01
FWIW the benchmark here doesn't indicate anything has changed in wcsaxes stuff but it can also be flaky. 🤷‍♀️ 

https://github.com/astropy/astropy/actions/runs/8911393283/job/24474385670?pr=16366