@astrofrog 2024-05-01
The improvements here are likely only for cases with slow coordinate transformations, which probably explains it.