@ayshih 2024-05-02
Amusingly, it's possible to make the performance gains from this PR arbitrarily large by increasing `conf.coordinate_range_samples` by orders of magnitude.  At `conf.coordinate_range_samples = 1000`, `find_coordinate_range()` can take ~1 second to run and dominates the total plotting run time, so this PR reduces the runtime by 75%.