@astrofrog 2024-07-08
@ayshih - do you think it's still worth pushing ahead with this? From your benchmark with the TETE overlay it looks like the results are mixed in that without the overlay it gets a little better but with the overlay it actually gets worse?