@ayshih 2024-07-08
Ha, whoops, there must be an error in my earlier post – I probably swapped the 480 ms and 540 ms – but I'll verify.  This PR is a strict improvement in performance.