@ayshih 2024-07-08
Here is a corrected timing comparison showing the improvement of this PR:

- Before this PR:
  - Without the overlay: 66 +/- 2 ms
  - With the overlay: 483 +/- 10 ms
- After this PR:
  - Without the overlay: 61 +/- 2 ms
  - With the overlay: 448 +/- 10 ms

I'll complain here that benchmarking plotting is a bit maddening because it's not uncommon for a run to take mysteriously long – e.g., ~10–20% longer than nominal – which completely throws off the average if outliers are not excluded...