@ayshih 2024-07-12
Now that #16694 has been merged, I have added a performance changelog entry here